Remove compilation error

+++ wire/src/serial.c
@@ -249,7 +249,7 @@ void serial_apply_change(void) {
     if (!isio.init) return;
     uint16_t len = strlen(isio.state.clabel.str);
     for (uint8_t i=0; i<CMD_LEN;i++){
-        if (isalpha(isio.state.clabel.str[i])){
+        if (isalpha((int)isio.state.clabel.str[i])){
             isio.state.clabel.str[i] &= 0xdf;
         }
     }
